[all] Fix invalid Memory access in MechanicalObject and Compliant_test (#1849)
* [Compliant_test] Add scope to WriteAccessor.

A write accessor is holding a pointer to the data. If the pointed data is deleted
before the accessor's then bad things will happen.
This is what s wrong with this test. It is fixed by adding scopes around
the accessors.

* [SofaBaseMechanicas] Fix wrong dimmension in MechanicalObject::applyScale

The existing code is making the assumption that the size is 3. Which is not true
for Vec1Type or Vec2Type as well as Rigid2Type.

* Update MechanicalObject.inl

* Update Assembly_test.cpp

* Fix

* FIXUP